requirements-dev.txt
amqp==1.4.5
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.18
blessings==1.5.1
boto==2.32.0
celery==3.1.13
coverage==3.7.1
dj-database-url==0.3.0
django-appconf==0.6
django-boto==0.3.1
django-celery==3.1.10
django-compressor==1.4
django-debug-toolbar==1.2.1
django-extensions==1.3.9
django-lockdown==1.0
django-nose==1.2
django-pylibmc==0.5.0
django-social-auth==0.7.28
django-storages==1.1.8
Django==1.6.5
djangorestframework==2.3.14
djorm-ext-pool==0.8.2
envdir==0.6.1
gunicorn==19.1.0
httplib2==0.9
kombu==3.0.21
memory-profiler==0.31
newrelic==2.22.1.20
nose-perfdump==1.5
nose-progressive==1.5.1
nose==1.3.3
oauth2==1.5.211
prettytable==0.7.2
psycopg2==2.5.3
pylibmc==1.3.0
pyparsing==2.0.2
python-openid==2.2.5
pytz==2014.4
raven==5.0.0
requests==2.3.0
simplejson==3.6.0
six==1.7.3
South==1.0
SQLAlchemy==0.9.7
sqlparse==0.1.11
stevedore==0.15
unicodecsv==0.9.4
virtualenv-clone==0.2.5
virtualenv==1.11.6
virtualenvwrapper==4.3.1
voluptuous==0.8.5
Werkzeug==0.9.6
wsgiref==0.1.2

